изменения: - фикс задержки на ~0.5 сек перед загрузкой главного меню - добавил колесо фортуны в Preloaded Assets, теперь при первом открытии колеса нет задержки - пофиксил все тени кнопок главного меню - target fps выставляется в зависимости от версии ОС, а также от того, в меню или в игре мы находимся - добавил id своих устройств в список тестовых id - добавил дополнительные кнопки запуска Profiler, Android Logcat и Color Palette в меню Tools - DebugConsoleController теперь GameObserver и по дефолту добавляется в список GameObserver-ов, можно писать отладочные методы команд прямо в GameObservable-классах и регистрировать команды через Notify - некоторые другие багфиксы и рефакторинг

+++ Assets/Scripts/Utils/GameUtils.cs
@@ -11,9 +11,9 @@ namespace Utils
             Vector2 gameViewSize = GetMainGameViewSize();
             aspectRatio = gameViewSize.x / gameViewSize.y;
 #else
-            aspectRatio = Screen.width / Screen.height;
+            aspectRatio = (float)Screen.width / (float)Screen.height;
 #endif
-            float vertExtent = Camera.main.orthographicSize;    
+            float vertExtent = Camera.main.orthographicSize;
             float horzExtent = vertExtent * aspectRatio;
             Vector3 size = new Vector3(horzExtent, vertExtent, 0);
             return new Bounds(Vector3.zero, size);
